+++ src/main/java/gift/auth/Login.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.auth;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
