+++ src/main/java/gift/auth/LoginService.java
@@ -1,9 +1,7 @@
 package gift.auth;
 
 import gift.exception.AuthException;
-import gift.dto.Login;
 import gift.model.member.LoginRepository;
-import gift.dto.Token;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
