+++ src/main/java/gift/auth/LoginUserArgumentResolver.java
@@ -7,7 +7,6 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import gift.dto.Login;
 import io.jsonwebtoken.Claims;
 
 @Component
