+++ src/main/java/gift/auth/Token.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.auth;
 
 public record Token(String token) {
